+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Fixes
 - Fixed inferring raw value type from inherited types for enums with no cases or with associated values
 - Fixed access level of protocol members
+- Fixes parsing indirect enum cases correctly even when inline documentation is used
 
 ## Internal changes
 - Removes manual parsing of `TypeName`, only explicit parser / configuration is now used